requirements-dev.txt
# Development / testing dependencies only.
# Install alongside the runtime deps:
#   pip install -r requirements.txt -r requirements-dev.txt

black>=24.0
ruff>=0.3.0
pytest>=8.0
pytest-asyncio>=0.23
homeassistant-stubs>=2024.1.0
